JavaScript click event is now configurable
Adds new configuration value to allow the JavaScript click event to be set
in your web.config file. By default, this will just close the DOM element,
and there is no reliance on jQuery, etc. However, it is now very easy to
add a $(this).slideUp() or any other desired method to remove the flash
message from the DOM.